@@ -4,31 +4,31 @@
     "keywords": [
         "fuelviews",
         "laravel",
-        "laravel-robots-txt",
-        "robots",
-        "php"
+        "robots-txt"
     ],
     "homepage": "https://github.com/fuelviews/laravel-robots-txt",
     "license": "MIT",
     "authors": [
         {
             "name": "Joshua Mitchener",
-            "email": "support@fuelviews.com",
             "homepage": "https://fuelviews.com",
             "role": "Developer"
         }
     ],
     "require": {
-        "php": "^8.2",
-        "spatie/laravel-package-tools": "^1.14.0"
+        "php": "^8.3",
+        "illuminate/contracts": "^10.0||^11.0||^12.0",
+        "spatie/laravel-package-tools": "^1.92"
     },
     "require-dev": {
-        "laravel/pint": "^1.0",
-        "nunomaduro/collision": "^7.8",
-        "orchestra/testbench": "^8.8",
-        "pestphp/pest": "^2.20",
-        "pestphp/pest-plugin-arch": "^2.5",
-        "pestphp/pest-plugin-laravel": "^2.0"
+        "laravel/pint": "^1.14",
+        "nunomaduro/collision": "^8.1.1||^7.10.0",
+        "orchestra/testbench": "^10.2||^9.0.0||^8.22.0",
+        "pestphp/pest": "^3.0||^2.34",
+        "pestphp/pest-plugin-arch": "^3.0||^2.7",
+        "pestphp/pest-plugin-laravel": "^3.2||^2.3",
+        "rector/rector": "^2.0",
+        "driftingly/rector-laravel": "^2.0"
     },
     "autoload": {
         "psr-4": {
