@@ -3,6 +3,7 @@
 namespace Fuelviews\RobotsTxt\Tests\Feature;
 
 use Fuelviews\RobotsTxt\Tests\TestCase;
+use PHPUnit\Framework\Attributes\Test;
 
 /**
  * Class RobotsTxtEnvironmentTest
@@ -14,46 +15,43 @@ class RobotsTxtEnvironmentTest extends TestCase
 {
     /**
      * Test if all paths are disallowed in non-production environments.
-     *
-     * @test
      */
-    public function it_disallows_in_local_environment()
+    #[Test]
+    public function it_disallows_in_local_environment(): void
     {
         $this->setTestConfigurations([
             'app.env' => 'local',
         ]);
 
-        $response = $this->get('robots.txt');
+        $testResponse = $this->get('robots.txt');
 
-        $response->assertStatus(200);
-        $response->assertSee('User-agent: *');
-        $response->assertSee('Disallow: /');
+        $testResponse->assertStatus(200);
+        $testResponse->assertSee('User-agent: *');
+        $testResponse->assertSee('Disallow: /');
     }
 
     /**
      * Test if all paths are disallowed in non-production environments.
-     *
-     * @test
      */
-    public function it_disallows_in_development_environment()
+    #[Test]
+    public function it_disallows_in_development_environment(): void
     {
         $this->setTestConfigurations([
             'app.env' => 'development',
         ]);
 
-        $response = $this->get('robots.txt');
+        $testResponse = $this->get('robots.txt');
 
-        $response->assertStatus(200);
-        $response->assertSee('User-agent: *');
-        $response->assertSee('Disallow: /');
+        $testResponse->assertStatus(200);
+        $testResponse->assertSee('User-agent: *');
+        $testResponse->assertSee('Disallow: /');
     }
 
     /**
      * Test if the robots.txt is generated correctly in the production environment.
-     *
-     * @test
      */
-    public function it_allows_in_production_environment()
+    #[Test]
+    public function it_allows_in_production_environment(): void
     {
         $this->setTestConfigurations([
             'app.env' => 'production',
@@ -61,22 +59,22 @@ class RobotsTxtEnvironmentTest extends TestCase
 
         $baseUrl = config('app.url');
 
-        $response = $this->get('robots.txt');
-        $response->assertStatus(200);
+        $testResponse = $this->get('robots.txt');
+        $testResponse->assertStatus(200);
 
         $userAgentsRules = config('robots-txt.user_agents');
         foreach ($userAgentsRules as $userAgent => $rules) {
             foreach ($rules as $ruleType => $paths) {
                 foreach ($paths as $path) {
-                    $response->assertSee("User-agent: $userAgent");
-                    $response->assertSee("$ruleType: $path");
+                    $testResponse->assertSee('User-agent: ' . $userAgent);
+                    $testResponse->assertSee(sprintf('%s: %s', $ruleType, $path));
                 }
             }
         }
 
         $sitemaps = config('robots-txt.sitemap');
-        foreach ($sitemaps as $sitemapPath) {
-            $response->assertSee('Sitemap: '.($baseUrl).'/'.$sitemapPath);
+        foreach ($sitemaps as $sitemap) {
+            $testResponse->assertSee('Sitemap: '.($baseUrl).'/'.$sitemap);
         }
     }
 }
